+++ src/HTML/Draw.php
@@ -660,9 +660,12 @@ class Draw extends HTMLMeta {
          * @see https://sortablejs.github.io/Sortable/
          * 
          * @since   LRS 3.15.0
+         * 
+         * @todo    Install via npm
+         * @todo    Try get composer to manage this so that it's not so dramatic
          */
         JS::script_module( "
-            import Sortable from './src/js/thirdparty/sortablejs/sortable.esm.js';
+            import Sortable from 'sortablejs';
             const el = document.getElementById('draggable_container{$id}');
             const sortable = Sortable.create(el, {
                 swap: true,